more precision boost

+++ src/Overlap.cpp
@@ -7,12 +7,12 @@ OverlapWindows::OverlapWindows(int32_t _nx, int32_t _ny, int32_t _ox, int32_t _o
 	ox = _ox;
 	oy = _oy;
 	size = nx*ny;
-	fWin1UVx = new float[nx];
-	fWin1UVxfirst = new float[nx];
-	fWin1UVxlast = new float[nx];
+	fWin1UVx = new double[nx];
+	fWin1UVxfirst = new double[nx];
+	fWin1UVxlast = new double[nx];
 	for (int32_t i = 0; i<ox; i++)
 	{
-		fWin1UVx[i] = static_cast <float> (cos(M_PI*(i - ox + 0.5f) / (ox * 2)));
+		fWin1UVx[i] = cos(M_PI*(i - ox + 0.5f) / (ox * 2));
 		fWin1UVx[i] = fWin1UVx[i] * fWin1UVx[i];// left window (rised cosine)
 		fWin1UVxfirst[i] = 1; // very first window
 		fWin1UVxlast[i] = fWin1UVx[i]; // very last
@@ -25,18 +25,18 @@ OverlapWindows::OverlapWindows(int32_t _nx, int32_t _ny, int32_t _ox, int32_t _o
 	}
 	for (int32_t i = nx - ox; i<nx; i++)
 	{
-		fWin1UVx[i] = static_cast <float> (cos(M_PI*(i - nx + ox + 0.5f) / (ox * 2)));
+		fWin1UVx[i] = cos(M_PI*(i - nx + ox + 0.5f) / (ox * 2));
 		fWin1UVx[i] = fWin1UVx[i] * fWin1UVx[i];// right window (falled cosine)
 		fWin1UVxfirst[i] = fWin1UVx[i]; // very first window
 		fWin1UVxlast[i] = 1; // very last
 	}
 
-	fWin1UVy = new float[ny];
-	fWin1UVyfirst = new float[ny];
-	fWin1UVylast = new float[ny];
+	fWin1UVy = new double[ny];
+	fWin1UVyfirst = new double[ny];
+	fWin1UVylast = new double[ny];
 	for (int32_t i = 0; i<oy; i++)
 	{
-		fWin1UVy[i] = static_cast <float> (cos(M_PI*(i - oy + 0.5f) / (oy * 2)));
+		fWin1UVy[i] = cos(M_PI*(i - oy + 0.5f) / (oy * 2));
 		fWin1UVy[i] = fWin1UVy[i] * fWin1UVy[i];// left window (rised cosine)
 		fWin1UVyfirst[i] = 1; // very first window
 		fWin1UVylast[i] = fWin1UVy[i]; // very last
@@ -49,7 +49,7 @@ OverlapWindows::OverlapWindows(int32_t _nx, int32_t _ny, int32_t _ox, int32_t _o
 	}
 	for (int32_t i = ny - oy; i<ny; i++)
 	{
-		fWin1UVy[i] = static_cast <float> (cos(M_PI*(i - ny + oy + 0.5f) / (oy * 2)));
+		fWin1UVy[i] = cos(M_PI*(i - ny + oy + 0.5f) / (oy * 2));
 		fWin1UVy[i] = fWin1UVy[i] * fWin1UVy[i];// right window (falled cosine)
 		fWin1UVyfirst[i] = fWin1UVy[i]; // very first window
 		fWin1UVylast[i] = 1; // very last
